Error update

+++ delete/delete.js
@@ -18,14 +18,29 @@ router.post("/", (req, res) => {
             console.log('Error in connection:- ', err)
         else {
             let db = conn.db("nodedb")
-            db.collection('products').deleteOne(obj, (err) => {
-                if (err)
-                    res.json({ 'delete': 'Error ' + err })
-                else {                    
-                    res.json({ 'delete': 'success' })
-                    conn.close()
+            db.collection('products').find({ p_id}).toArray(
+                (err,recodsArray)=>{
+                    if(err)
+                    console.log('Error:- ' +err)
+                    else{
+                        if(recodsArray.length >0){
+                            db.collection('products').deleteOne(obj, (err) => {
+                                if (err)
+                                    res.json({ 'delete': 'Error ' + err })
+                                else {                    
+                                    res.json({ 'delete': 'success' })
+                                    conn.close()
+                                }
+                            })
+                        }
+                        else{
+                            res.json('Record not found')
+                            conn.close()
+                        }
+                    }
                 }
-            })
+            )
+            
         }
     })
 })
